+++ packages/smarthr-ui/src/components/Dropdown/SortDropdown/SortDropdown.tsx
@@ -1,19 +1,26 @@
 'use client'
 
 import {
+  type ChangeEvent,
   type ComponentPropsWithRef,
   type FC,
   type FormEvent,
+  type MouseEvent,
   type MouseEventHandler,
   type OptionHTMLAttributes,
   type PropsWithChildren,
   type ReactNode,
   memo,
+  useMemo,
+  useState,
 } from 'react'
 
+import { useLatest } from '../../../hooks/useLatest'
+import { useIntl } from '../../../intl'
 import { Button } from '../../Button'
 import { Fieldset } from '../../Fieldset'
 import { FormControl } from '../../FormControl'
+import { FaArrowDownWideShortIcon, FaArrowUpWideShortIcon } from '../../Icon'
 import { Cluster, Stack } from '../../Layout'
 import { RadioButton } from '../../RadioButton'
 import { Select } from '../../Select'
@@ -21,8 +28,6 @@ import { Dropdown } from '../Dropdown'
 import { DropdownCloser } from '../DropdownCloser'
 import { DropdownContent } from '../DropdownContent'
 import { DropdownTrigger } from '../DropdownTrigger'
-
-import { useSortDropdown } from './useSortDropdown'
 
 type SortFieldType = {
   value: string
@@ -69,29 +74,128 @@ export const SortDropdown: FC<Props> = ({
   onCancel,
   ...rest
 }) => {
-  const {
-    texts,
-    SortIcon,
-    innerValues: { innerFields, innerCheckedOrder },
-    handler,
-  } = useSortDropdown({
-    sortFields,
-    defaultOrder,
+  const { localize } = useIntl()
+
+  const texts = useMemo(
+    () => ({
+      sortFieldLabel:
+        sortFieldLabel ||
+        localize({
+          id: 'smarthr-ui/SortDropdown/sortFieldLabel',
+          defaultText: '並べ替え項目',
+        }),
+      sortOrderLegend:
+        sortOrderLegend ||
+        localize({
+          id: 'smarthr-ui/SortDropdown/sortOrderLegend',
+          defaultText: '並び順',
+        }),
+      ascLabel:
+        ascLabel ||
+        localize({
+          id: 'smarthr-ui/SortDropdown/ascLabel',
+          defaultText: '昇順',
+        }),
+      descLabel:
+        descLabel ||
+        localize({
+          id: 'smarthr-ui/SortDropdown/descLabel',
+          defaultText: '降順',
+        }),
+      applyText:
+        applyText ||
+        localize({
+          id: 'smarthr-ui/SortDropdown/applyText',
+          defaultText: '適用',
+        }),
+      cancelText:
+        cancelText ||
+        localize({
+          id: 'smarthr-ui/SortDropdown/cancelText',
+          defaultText: 'キャンセル',
+        }),
+    }),
+    [sortFieldLabel, sortOrderLegend, ascLabel, descLabel, applyText, cancelText, localize],
+  )
+
+  const [defaultFieldLabel] = useState(
+    () => (sortFields.find((field) => field.selected) || sortFields[0])?.label || '',
+  )
+
+  // 外向きの値
+  const [selectedLabel, setSelectedLabel] = useState<string>(defaultFieldLabel)
+  const [checkedOrder, setCheckedOrder] = useState<Props['defaultOrder']>(defaultOrder)
+
+  // 内部的な値
+  const [innerFields, setInnerFields] = useState<Props['sortFields']>(sortFields)
+  const [innerSelectedField, setInnerSelectedField] = useState<string>(defaultFieldLabel)
+  const [innerCheckedOrder, setCheckedInnerOrder] = useState<Props['defaultOrder']>(defaultOrder)
+
+  const hasOnCancel = !!onCancel
+
+  const latest = useLatest({
+    innerCheckedOrder,
+    innerFields,
+    innerSelectedField,
     onApply,
     onCancel,
-    sortFieldLabel,
-    sortOrderLegend,
-    ascLabel,
-    descLabel,
-    applyText,
-    cancelText,
   })
+
+  const handler = useMemo(
+    () => ({
+      change: (e: ChangeEvent<HTMLSelectElement>) => {
+        const select = e.currentTarget
+        const newLabel = select.options[select.selectedIndex].label
+
+        setInnerFields((currentFields) =>
+          currentFields.map((field) => {
+            if (field.label === newLabel) {
+              if (!field.selected) {
+                return {
+                  ...field,
+                  selected: true,
+                }
+              }
+            } else if (field.selected) {
+              return {
+                ...field,
+                selected: false,
+              }
+            }
+
+            return field
+          }),
+        )
+        setInnerSelectedField(newLabel)
+      },
+      apply: () => {
+        setSelectedLabel(latest.innerSelectedField)
+        setCheckedOrder(latest.innerCheckedOrder)
+        latest.onApply({
+          field: latest.innerSelectedField || '',
+          order: latest.innerCheckedOrder,
+          newfields: latest.innerFields,
+        })
+      },
+      cancel: hasOnCancel
+        ? (e: MouseEvent<HTMLButtonElement>) => {
+            latest.onCancel!(e)
+          }
+        : undefined,
+      changeSortOrderRadio: (e: ChangeEvent<HTMLInputElement>) => {
+        setCheckedInnerOrder(e.currentTarget.value as Props['defaultOrder'])
+      },
+    }),
+    [hasOnCancel, latest],
+  )
+
+  const SortIcon = checkedOrder === 'asc' ? FaArrowUpWideShortIcon : FaArrowDownWideShortIcon
 
   return (
     <Dropdown>
       <DropdownTrigger>
         <Button {...rest} suffix={<SortIcon />}>
-          {texts.triggerLabel}
+          {`${selectedLabel}（${checkedOrder === 'asc' ? texts.ascLabel : texts.descLabel}）`}
         </Button>
       </DropdownTrigger>
       <DropdownContent controllable>
